Data extraction to objects

+++ AnimationOpenGL/src/Animation.h
@@ -25,7 +25,7 @@ public:
 
 	~Animation()
 	{
-		for (int i = 0; i < m_keyFrames.size(); i++)
+		for (size_t i = 0; i < m_keyFrames.size(); i++)
 		{
 			delete m_keyFrames[i];
 		}
