+++ AnimationOpenGL/src/AnimationData.h
@@ -17,7 +17,7 @@ public:
 
 	~AnimationData()
 	{
-		for (int i = 0; i < keyFrames.size(); i++)
+		for (size_t i = 0; i < keyFrames.size(); i++)
 		{
 			delete keyFrames[i];
 		}
